requirements.txt
Flask==2.0.1
selenium==4.4.0
python-dateutil==2.8.2